Initial geo-traits implementation
