+++ geo-traits/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "geo-traits"
+version = "0.1.0"
+license = "MIT OR Apache-2.0"
+repository = "https://github.com/georust/geo"
+documentation = "https://docs.rs/geo-traits/"
+readme = "../README.md"
+keywords = ["gis", "geo", "geography", "geospatial"]
+description = "Geospatial traits"
+rust-version = "1.63"
+edition = "2021"
+
+[features]
+
+[dependencies]
+geo-types = "0.7.9"
+
+[dev-dependencies]
+approx = ">= 0.4.0, < 0.6.0"
